some changes in job banner
